### 2023-01-26 #102 @lucaferranti
@vchuravy if you havent found any reviewers, I could review this, I have some working experience with satellite positioning and receivers architecture, as well as Julia working experience ofc
