### 2023-01-26 #102 @ozmaden
@vchuravy I will fix the DOI asap, seems like I just need to remove the 'https://doi.org/' prefix.
 I do not have specific reviewers in mind. If @lucaferranti would like to review my work, I would be very happy to get his feedback
