### 2023-04-21 #128 @whedon
[REVIEW]: Tracking.jl: Accelerating multi-antenna GNSS receivers with CUDA
**Submitting author:** @ozmaden (<a href="http://orcid.org/0000-0003-0614-3988">Can Özmaden</a>)
**Repository:** <a href="https://github.com/ozmaden/GPUAcceleratedTracking" target ="_blank">https://github.com/ozmaden/GPUAcceleratedTracking</a>
**Version:**
