### 2023-04-24 ozmaden/GPUAcceleratedTracking#5 @coezmaden
Hi @maleadt! Thank you very much for your thorough and insightful review. I sincerely appreciate your effort and time invested to review my paper. At the moment I am on a vacation until the 2nd week of May, so please excuse the late response. I will get back later in May with the suggested improvements. Thanks again!
