### 2023-05-19 ozmaden/GPUAcceleratedTracking#5 @zsoerenm
@coezmaden I made quite a big restructure to Tracking.jl:
https://github.com/JuliaGNSS/Tracking.jl/pull/43
With this pull request you are able to track multiple satellites coherently.
I adjusted the GPU implementation:
https://github.com/JuliaGNSS/Tracking.jl/blob/master/src/gpu_downconvert_and_correlate.jl
I have not optimized it to the last, but it may stand as a good baseline.
On my 10 year old computer the GPU version is about half as slow as the CPU version.
The implementation could in theory be parallelized over multiple number of satellites, that wasn't possible beforehand.
The kernel is quite simple:
https://github.com/JuliaGNSS/Tracking.jl/blob/master/src/gpu_downconvert_and_correlate.jl#L127
The kernel only downconverts and decodes the PRN.
The correlation (the sum over the samples) is done afterwards: https://github.com/JuliaGNSS/Tracking.jl/blob/master/src/gpu_downconvert_and_correlate.jl#L191

Have a look at this test, if you'd like to know how to use the implementation:
https://github.com/JuliaGNSS/Tracking.jl/blob/master/test/track.jl#L390
