### 2023-11-06 #128 @fcdimitr
:wave: @coezmaden: I see that @maleadt has posted his review in a separate issue (https://github.com/coezmaden/GPUAcceleratedTracking/issues/5). Could you update us on your efforts to address the comments and revise the manuscript?

:wave: @lucaferranti: Could you give us an update on how your review is going?

Thank you!
