### 2023-11-17 #128 @coezmaden
> 👋 @coezmaden: Could you please update us on the progress of addressing the reviewer's comments? I enabled the flag `author-action-required` until we hear back from you. If we do not receive a response within 25 days, the issue will be closed, and you will have to start a new submission from scratch. Thank you!

Sorry for the late reply and inactivity. The original hardware and the funding for the project behind this submission has been finalized after the review came. I am afraid I cannot provide any viable update to the paper due to time and the lack of original hardware reasons. Apologies for the caused inconvenience, but I must retract this submission.
